More work on flash support
More useful flash abstraction (flash_hal.c), small emulator bugfix,
and get the loadstream decoder building with sdcc.

+++ emulator/cube/flash.c
@@ -28,7 +28,7 @@
 #define FLUSH_TIME_US  100000
 
 struct cmd_state {
-    uint16_t addr;
+    uint32_t addr;
     uint8_t data;
 };
